Move EntryTypeConfiguration module out of Configuration
Part of top level public API. Also move entry type specific
configuration class out of plugin classes.

Extend docs.
