+++ entry_types/paged/lib/pageflow_paged/configuration.rb
@@ -1,0 +1,6 @@
+module PageflowPaged
+  # Configuration options for paged entry type
+  class Configuration
+    include Pageflow::EntryTypeConfiguration
+  end
+end
