+++ entry_types/paged/lib/pageflow_paged/plugin.rb
@@ -4,9 +4,5 @@ module PageflowPaged
     def configure(config)
       config.entry_types.register(PageflowPaged.entry_type)
     end
-
-    class PagedConfiguration
-      include Pageflow::Configuration::EntryTypeConfiguration
-    end
   end
 end
