+++ entry_types/scrolled/lib/pageflow_scrolled.rb
@@ -10,7 +10,7 @@ module PageflowScrolled
     def entry_type
       Pageflow::EntryType.new(name: 'scrolled',
                               frontend_app: PageflowScrolled::EntriesController.action(:show),
-                              configuration: Plugin::ScrolledConfiguration,
+                              configuration: PageflowScrolled::Configuration,
                               editor_fragment_renderer: editor_fragment_renderer,
                               editor_app: PageflowScrolled::Engine)
     end
