+++ entry_types/scrolled/lib/pageflow_scrolled/configuration.rb
@@ -1,0 +1,6 @@
+module PageflowScrolled
+  # Configuration options for scrolled entry type
+  class Configuration
+    include Pageflow::EntryTypeConfiguration
+  end
+end
