+++ entry_types/scrolled/lib/pageflow_scrolled/plugin.rb
@@ -4,9 +4,5 @@ module PageflowScrolled
     def configure(config)
       config.entry_types.register(PageflowScrolled.entry_type)
     end
-
-    class ScrolledConfiguration
-      include Pageflow::Configuration::EntryTypeConfiguration
-    end
   end
 end
